requirements.txt
tensorflow-cpu
opencv-python
matplotlib
pillow
numpy
streamlit
pandas
